[Tool] Support Darwin BE bootstrap flow (#70230)

+++ build.sh
@@ -33,6 +33,11 @@
 # compiled and installed correctly.
 ##############################################################
 startTime=$(date +%s)
+ROOT=`dirname "$0"`
+ROOT=`cd "$ROOT"; pwd`
+export STARROCKS_HOME=${ROOT}
+
+. "${STARROCKS_HOME}/build-support/build_helpers.sh"
 
 format_time() {
     local epoch="$1"
@@ -42,11 +47,11 @@ format_time() {
         date -d "@${epoch}" '+%Y-%m-%d %H:%M:%S'
     fi
 }
-ROOT=`dirname "$0"`
-ROOT=`cd "$ROOT"; pwd`
 MACHINE_TYPE=$(uname -m)
 
-export STARROCKS_HOME=${ROOT}
+is_aarch64_host() {
+    [[ "${MACHINE_TYPE}" == "aarch64" || "${MACHINE_TYPE}" == "arm64" ]]
+}
 
 if [ -z $BUILD_TYPE ]; then
     export BUILD_TYPE=Release
@@ -74,17 +79,22 @@ if [ -z $STARROCKS_COMMIT_HASH ] ; then
 fi
 
 set -eo pipefail
-. ${STARROCKS_HOME}/env.sh
+if starrocks_is_darwin; then
+    starrocks_setup_darwin_build_env
+else
+    . ${STARROCKS_HOME}/env.sh
+fi
 
-if [[ $OSTYPE == darwin* ]] ; then
-    PARALLEL=$(sysctl -n hw.ncpu)
-    # We know for sure that build-thirdparty.sh will fail on darwin platform, so just skip the step.
+if starrocks_is_darwin ; then
+    PARALLEL=$(starrocks_detect_parallelism)
+    # Darwin thirdparty is prepared separately and validated before BE configure.
 else
     if [[ ! -f ${STARROCKS_THIRDPARTY}/installed/llvm/lib/libLLVMInstCombine.a ]]; then
         echo "Thirdparty libraries need to be build ..."
         ${STARROCKS_THIRDPARTY}/build-thirdparty.sh
     fi
-    PARALLEL=$[$(nproc)/4+1]
+    host_parallelism=$(starrocks_detect_parallelism)
+    PARALLEL=$[$host_parallelism/4+1]
 fi
 
 # Check args
@@ -144,22 +154,7 @@ Usage: $0 <options>
   exit 1
 }
 
-GETOPT_BIN="getopt"
-if [[ "$OSTYPE" == darwin* ]]; then
-    # Try to detect gnu-getopt path dynamically
-    if command -v brew &> /dev/null; then
-        GNU_GETOPT_PREFIX=$(brew --prefix gnu-getopt 2>/dev/null)
-        if [[ -n "${GNU_GETOPT_PREFIX}" ]] && [[ -x "${GNU_GETOPT_PREFIX}/bin/getopt" ]]; then
-            GETOPT_BIN="${GNU_GETOPT_PREFIX}/bin/getopt"
-        else
-            echo "gnu-getopt is required on macOS. Please install it with 'brew install gnu-getopt'."
-            exit 1
-        fi
-    else
-        echo "Homebrew is required on macOS to install gnu-getopt. Please install Homebrew first."
-        exit 1
-    fi
-fi
+GETOPT_BIN="$(starrocks_resolve_getopt_bin)" || exit 1
 
 OPTS=$(${GETOPT_BIN} \
   -n $0 \
@@ -210,12 +205,20 @@ WITH_GCOV=OFF
 WITH_BENCH=OFF
 WITH_CLANG_TIDY=OFF
 WITH_COMPRESS=ON
-WITH_STARCACHE=ON
+if starrocks_is_darwin; then
+    WITH_STARCACHE=OFF
+else
+    WITH_STARCACHE=ON
+fi
 WITH_PCH=ON
 USE_STAROS=OFF
 BUILD_JAVA_EXT=ON
 OUTPUT_COMPILE_TIME=OFF
-WITH_TENANN=ON
+if starrocks_is_darwin; then
+    WITH_TENANN=OFF
+else
+    WITH_TENANN=ON
+fi
 CONFIGURE_ONLY=OFF
 WITH_RELATIVE_SRC_PATH=ON
 ENABLE_MULTI_DYNAMIC_LIBS=OFF
@@ -249,7 +252,11 @@ if [[ -z ${USE_BMI_2} ]]; then
     USE_BMI_2=ON
 fi
 if [[ -z ${ENABLE_JIT} ]]; then
-    ENABLE_JIT=ON
+    if starrocks_is_darwin; then
+        ENABLE_JIT=OFF
+    else
+        ENABLE_JIT=ON
+    fi
 fi
 if [[ -z ${DISABLE_JAVA_CHECK_STYLE} ]]; then
     DISABLE_JAVA_CHECK_STYLE=OFF
@@ -368,6 +375,10 @@ if [ ${BUILD_FORMAT_LIB} -eq 1 ]; then
     BUILD_JAVA_EXT=OFF
 fi
 
+if starrocks_is_darwin && { [ ${BUILD_BE} -eq 1 ] || [ ${BUILD_FORMAT_LIB} -eq 1 ]; }; then
+    starrocks_validate_darwin_thirdparty
+fi
+
 echo "Get params:
     BUILD_BE                    -- $BUILD_BE
     BUILD_FORMAT_LIB            -- $BUILD_FORMAT_LIB
@@ -433,10 +444,18 @@ if [ ${BUILD_BE} -eq 1 ] || [ ${BUILD_FORMAT_LIB} -eq 1 ] ; then
 fi
 cd ${STARROCKS_HOME}
 
-if [[ "${MACHINE_TYPE}" == "aarch64" ]]; then
-    export LIBRARY_PATH=${JAVA_HOME}/jre/lib/aarch64/server/
+JAVA_LIBRARY_PATH=""
+if starrocks_is_darwin; then
+    if [[ -d "${JAVA_HOME}/lib/server" ]]; then
+        JAVA_LIBRARY_PATH="${JAVA_HOME}/lib/server:${JAVA_HOME}/lib"
+    fi
+elif is_aarch64_host; then
+    JAVA_LIBRARY_PATH=${JAVA_HOME}/jre/lib/aarch64/server/
 else
-    export LIBRARY_PATH=${JAVA_HOME}/jre/lib/amd64/server/
+    JAVA_LIBRARY_PATH=${JAVA_HOME}/jre/lib/amd64/server/
+fi
+if [[ -n "${JAVA_LIBRARY_PATH}" ]]; then
+    export LIBRARY_PATH="${JAVA_LIBRARY_PATH}${LIBRARY_PATH:+:${LIBRARY_PATH}}"
 fi
 
 addon_mvn_opts=""
@@ -648,7 +667,7 @@ if [ ${BUILD_FORMAT_LIB} -eq 1 ]; then
     cp -r ${STARROCKS_HOME}/be/output/format-lib/* ${STARROCKS_OUTPUT}/format-lib/
     # format $BUILD_TYPE to lower case
     ibuildtype=`echo ${BUILD_TYPE} | tr 'A-Z' 'a-z'`
-    if [ "${ibuildtype}" == "release" ] ; then
+    if [ "${ibuildtype}" == "release" ] && ! starrocks_is_darwin ; then
         pushd ${STARROCKS_OUTPUT}/format-lib/ &>/dev/null
         FORMAT_LIB=libstarrocks_format.so
         FORMAT_LIB_DEBUGINFO=libstarrocks_format.debuginfo
@@ -687,13 +706,46 @@ if [ ${BUILD_BE} -eq 1 ]; then
         cp -r -p ${STARROCKS_HOME}/be/output/conf/asan_suppressions.conf ${STARROCKS_OUTPUT}/be/conf/
     fi
     cp -r -p ${STARROCKS_HOME}/be/output/lib/starrocks_be ${STARROCKS_OUTPUT}/be/lib/
-    cp -r -p ${STARROCKS_HOME}/be/output/lib/*.so ${STARROCKS_HOME}/be/output/lib/*.so.* ${STARROCKS_OUTPUT}/be/lib/ 2>/dev/null || true
-    mv ${STARROCKS_OUTPUT}/be/lib/libmockjvm.so ${STARROCKS_OUTPUT}/be/lib/libjvm.so
-    cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc/bin/jeprof ${STARROCKS_OUTPUT}/be/bin
+    shopt -s nullglob
+    be_shared_libs=(${STARROCKS_HOME}/be/output/lib/*.so ${STARROCKS_HOME}/be/output/lib/*.so.*)
+    if starrocks_is_darwin; then
+        be_shared_libs+=(${STARROCKS_HOME}/be/output/lib/*.dylib ${STARROCKS_HOME}/be/output/lib/*.dylib.*)
+    fi
+    if (( ${#be_shared_libs[@]} > 0 )); then
+        cp -r -p "${be_shared_libs[@]}" ${STARROCKS_OUTPUT}/be/lib/
+    fi
+    if starrocks_is_darwin; then
+        if [[ -f ${STARROCKS_OUTPUT}/be/lib/libmockjvm.dylib && ! -e ${STARROCKS_OUTPUT}/be/lib/libjvm.dylib ]]; then
+            cp -p ${STARROCKS_OUTPUT}/be/lib/libmockjvm.dylib ${STARROCKS_OUTPUT}/be/lib/libjvm.dylib
+        fi
+    elif [[ -f ${STARROCKS_OUTPUT}/be/lib/libmockjvm.so ]]; then
+        mv ${STARROCKS_OUTPUT}/be/lib/libmockjvm.so ${STARROCKS_OUTPUT}/be/lib/libjvm.so
+    fi
+    if [[ -f ${STARROCKS_THIRDPARTY}/installed/jemalloc/bin/jeprof ]]; then
+        cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc/bin/jeprof ${STARROCKS_OUTPUT}/be/bin
+    fi
     mkdir -p ${STARROCKS_OUTPUT}/be/lib/jemalloc
-    cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc/lib-shared/libjemalloc.so.2 ${STARROCKS_OUTPUT}/be/lib/jemalloc/libjemalloc.so.2
-    mkdir -p ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg
-    cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc-debug/lib/libjemalloc.so.2 ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg/libjemalloc.so.2
+    if starrocks_is_darwin; then
+        jemalloc_release_libs=(${STARROCKS_THIRDPARTY}/installed/jemalloc/lib-shared/libjemalloc*.dylib ${STARROCKS_THIRDPARTY}/installed/jemalloc/lib-shared/libjemalloc*.dylib.*)
+        if (( ${#jemalloc_release_libs[@]} > 0 )); then
+            cp -r -p "${jemalloc_release_libs[@]}" ${STARROCKS_OUTPUT}/be/lib/jemalloc/
+        fi
+        jemalloc_debug_libs=(${STARROCKS_THIRDPARTY}/installed/jemalloc-debug/lib/libjemalloc*.dylib ${STARROCKS_THIRDPARTY}/installed/jemalloc-debug/lib/libjemalloc*.dylib.*)
+        if (( ${#jemalloc_debug_libs[@]} > 0 )); then
+            mkdir -p ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg
+            cp -r -p "${jemalloc_debug_libs[@]}" ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg/
+        fi
+    else
+        cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc/lib-shared/libjemalloc.so.2 ${STARROCKS_OUTPUT}/be/lib/jemalloc/libjemalloc.so.2
+        mkdir -p ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg
+        cp -r -p ${STARROCKS_THIRDPARTY}/installed/jemalloc-debug/lib/libjemalloc.so.2 ${STARROCKS_OUTPUT}/be/lib/jemalloc-dbg/libjemalloc.so.2
+    fi
+    if starrocks_is_darwin; then
+        starrocks_write_output_be_host_dylib_manifest "${STARROCKS_OUTPUT}/be"
+        echo "Warning: macOS output/be depends on host dylibs not bundled in output/be."
+        echo "See ${STARROCKS_OUTPUT}/be/HOST_DYLIB_DEPENDENCIES.txt for the required host library paths."
+    fi
+    shopt -u nullglob
 
     # Copy pprof and FlameGraph tools
     if [ -d "${STARROCKS_THIRDPARTY}/installed/flamegraph" ]; then
@@ -703,7 +755,7 @@ if [ ${BUILD_BE} -eq 1 ]; then
 
     # format $BUILD_TYPE to lower case
     ibuildtype=`echo ${BUILD_TYPE} | tr 'A-Z' 'a-z'`
-    if [ "${ibuildtype}" == "release" ] ; then
+    if [ "${ibuildtype}" == "release" ] && ! starrocks_is_darwin ; then
         pushd ${STARROCKS_OUTPUT}/be/lib/ &>/dev/null
         BE_BIN=starrocks_be
         BE_BIN_DEBUGINFO=starrocks_be.debuginfo
